Improve enum naming for '-' and '_' characters
